Movement
Applied the Display() philosophy to Movement(). Currently does not take input but entities can be moved by Movement(). Additionally, entities now abide by speed, velocity, and acceleration when moving. This adds a nice look to the way things move (not just immediate stop and start) and adds a nice feel to player movement. This "game feel" was accomplished to great effect in 'kawaii' but this is a more powerful solution.